+++ src/vite/transformRoute.ts
@@ -1,18 +1,7 @@
 import MagicString from "magic-string";
 import { parse } from "@typescript-eslint/typescript-estree";
 import { walk } from "./walk";
-
-const isErrorBoundaryExport = (node: any) => {
-  return (
-    node.type === "ExportNamedDeclaration" &&
-    node.declaration &&
-    ((node.declaration.type === "FunctionDeclaration" &&
-      node.declaration.id.name === "ErrorBoundary") ||
-      (node.declaration.type === "VariableDeclaration" &&
-        node.declaration.declarations.length === 1 &&
-        node.declaration.declarations[0].id.name === "ErrorBoundary"))
-  );
-};
+import { isNamedExport } from "./isNamedExport";
 
 export function transformRoute({
   code,
@@ -33,20 +22,53 @@ export function transformRoute({
   const magicString = new MagicString(code, { filename: id });
 
   // prettier-ignore
-  const importStr = 'import { withErrorBoundary } from "@metronome-sh/dev-error-boundary/react";\n';
-  const stylesStr = 'import "@metronome-sh/dev-error-boundary/styles";\n';
+  const importStr = 'import { withErrorBoundary, withErrorBoundaryLinks } from "@metronome-sh/dev-error-boundary/react";\n';
+  // prettier-ignore
+  const stylesStr = 'import devErrorBoundaryStyles from "@metronome-sh/dev-error-boundary/styles?url";\n';
 
   const appDirectoryStr = JSON.stringify(appDirectory);
 
   if (!code.includes(importStr)) magicString.prepend(importStr);
-
-  const isRoot = id.match(/\/root\.[jt]sx$/);
-  if (isRoot && !code.includes(stylesStr)) magicString.prepend(stylesStr);
+  if (!code.includes(stylesStr)) magicString.prepend(stylesStr);
 
   let errorBoundaryFound = false;
+  let linksExportFound = false;
 
   walk(ast, (node) => {
-    if (isErrorBoundaryExport(node)) {
+    if (isNamedExport(node, "links")) {
+      linksExportFound = true;
+
+      const [start, end] = node.declaration.range;
+
+      if (node.declaration.type === "FunctionDeclaration") {
+        const declarationCode = code.substring(start, end).replace(/;$/g, "");
+
+        const replacementCode = `const links = withErrorBoundaryLinks(devErrorBoundaryStyles, ${declarationCode})`;
+
+        magicString.overwrite(start, end, replacementCode);
+
+        return;
+      }
+
+      if (node.declaration.type === "VariableDeclaration") {
+        const declarationCode = code
+          .substring(start, end)
+          .replace(/^.*links\s*(?::[^=]+)?=/g, "")
+          .replace(/;$/g, "");
+
+        const replacementCode = `const links = withErrorBoundaryLinks(devErrorBoundaryStyles, ${declarationCode})`;
+
+        magicString.overwrite(start, end, replacementCode);
+        return;
+      }
+
+      throw new Error(
+        "[dev-error-boundary] Unhandled links export type:",
+        node.declaration.type
+      );
+    }
+
+    if (isNamedExport(node, "ErrorBoundary")) {
       errorBoundaryFound = true;
 
       const [start, end] = node.declaration.range;
@@ -80,11 +102,11 @@ export function transformRoute({
     }
   });
 
-  if (!errorBoundaryFound) {
-    magicString.append(
-      `\nexport const ErrorBoundary = withErrorBoundary(${appDirectoryStr});\n`
-    );
-  }
+  // prettier-ignore
+  if (!errorBoundaryFound) magicString.append(`\nexport const ErrorBoundary = withErrorBoundary(${appDirectoryStr});\n`)
+
+  // prettier-ignore
+  if (!linksExportFound) magicString.append(`\nexport const links = () => [{ rel: "stylesheet", href: devErrorBoundaryStyles }];\n`);
 
   return {
     code: magicString.toString(),